CI: Build with Mono 6.12.0.144 + other env updates
Update env components to these versions:

- Python 3.9
- Pin NDK to 21.4.7075529 as done by Godot

Not updating Android API level to 30 yet as some feature work is needed in
Godot to support new requirements.

Mono 6.12.0.144 is still a Preview release, but contrarily to 6.12.0.122,
it has a successful Windows build. It also theoretically finalizes support
for Apple M1, so we can attempt using it to build Mono for that arch.